Add fabricEnabled config, update epi chart version to 0.6.14
A new configuration "fabricEnabled" has been introduced in the values.yaml file. This configuration enables or disables certain features based on its value. Alongside this, the epi Helm chart version has been updated to 0.6.14 and this update is also mirrored in the README file's installation instructions as well as in the deployment.yaml file.