Add ingress configurations for leaflet-reader
This commit includes enabling the leafletReader object in the values.yaml file and defining related ingress configurations in the leaflet-reader-ingress.yaml file. It also adds a condition to the Service and Deployment templates to check if leafletReader is enabled before proceeding. Last but not least, it updates the service port number in the leaflet-reader-service.yaml file.